requirements.txt
fastapi
uvicorn
python-multipart
langdetect
markdown2
beautifulsoup4
requests
pdfminer.six
python-docx
numpy
